requirements.txt
fastapi
uvicorn
transformers
sentencepiece
torch
python-docx
pypdf2
nltk
spacy
datasets
imblearn
sentence-transformers
